pyproject.toml
[project]
name = "django-better-errors"
version = "0.1.0"
description = "An implementation of the RFC 9457"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
    {name = "Ferran Jovell", email = "ferran.jovell+gh@gmail.com"}
]
keywords = ["django"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
    "Environment :: Web Environment",
    "Framework :: Django"
]
 dependencies = [
    "Django>=3.2.14",
    "djangorestframework",
 ]

[project.urls]
homepage = "https://github.com/mrswats/django-better-errors"
"Bug Tracker" = "https://github.com/mrswats/django-better-errors/issues"

[project.optional-dependencies]
test = [
    "django-stubs[compatible-mypy]",
    "djangorestframework-stubs[compatible-mypy]",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "pytest-randomly",
    "pytest-xdist",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["django_better_errors"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "testing.settings"

[tool.black]
line-length = 99
target_version = ["py38"]
skip-string-normalization = false

[tool.mypy]
plugins = ["mypy_django_plugin.main","mypy_drf_plugin.main"]
files = "django_better_errors"
namespace_packages = true
explicit_package_bases = true
show_error_codes = true
strict = true
enable_error_code = [
    "ignore-without-code",
    "redundant-expr",
    "truthy-bool",
]
exclude = [
   "tests/*.py",
   "testing/*.py",
]

[tool.django-stubs]
django_settings_module = "testing.settings"

[tool.coverage.run]
source = ["."]
branch = true
omit = [
    "manage.py",
    "**/tests/*",
    "**/migrations/*.py",
]

[tool.coverage.report]
skip_empty = true
fail_under = 90

[tool.coverage.html]
directory = "artifacts/htmlcov/"
